pyproject.toml
[project]
name = "anniegodfather"
version = "0.1.0"
description = "Telegram bot for Annie project"
authors = [
    {name = "Sergey Voronov",email = "graf.voronov@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.2"
dependencies = [
    "aiogram (>=3.17.0,<4.0.0)",
    "telethon (>=1.39.0,<2.0.0)",
    "grpcio (>=1.71.0,<2.0.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "redis (>=7.0.1,<8.0.0)",
    "jwt (>=1.4.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.app.dependencies]
pydantic = "^2.10.6"
dynaconf = "^3.2.7"
python-dotenv = "^1.0.1"


[tool.poetry.dependencies]
